--- Makefile
@@ -1,6 +1,0 @@
-run:
-	@go run main.go
-build:
-	@go build
-
-.PHONY: all build deps lint test
