+++ api/middleware/cors.go
