--- index.html
@@ -1,9 +1,0 @@
-<DOCTYPE html> 
-<head>
-    <title>asd</title>
-</head>
-
-<body>
-</body>
-
-</html>
