--- picori/user.go
@@ -1,14 +1,0 @@
-package picori
-
-import "time"
-
-type User struct {
-    Id string
-    Username string
-    Created time.Time
-}
-
-type UserFilter struct {
-    Id string
-    Username string
-}
