+++ storage/storage.go
@@ -1,10 +1,10 @@
 package storage
 
 import (
-	"github.com/gabriel-wer/picori/picori"
+	"github.com/gabriel-wer/picori"
 )
 
 type Storage interface {
-    GetURL(*picori.URL) error
-    InsertURL(picori.URL) error
+	GetURL(*picori.URL) error
+	InsertURL(picori.URL) error
 }
