+++ user.go
@@ -1,0 +1,14 @@
+package picori
+
+import "time"
+
+type User struct {
+	Id       string
+	Username string
+	Created  time.Time
+}
+
+type UserFilter struct {
+	Id       string
+	Username string
+}
