+++ util/errors.go
@@ -1,17 +1,17 @@
 package errors
 
 import (
+	"encoding/json"
 	"net/http"
-    "encoding/json"
 
-	"github.com/gabriel-wer/picori/picori"
+	"github.com/gabriel-wer/picori"
 )
 
 func JsonData(w http.ResponseWriter, url picori.URL) []byte {
-    jsonData, err := json.Marshal(url)
-    if err != nil{ 
-        w.Write([]byte("Cannot marshall JSON Data"))
-        http.Error(w, err.Error(), http.StatusInternalServerError)
-    }
-    return jsonData
+	jsonData, err := json.Marshal(url)
+	if err != nil {
+		w.Write([]byte("Cannot marshall JSON Data"))
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+	}
+	return jsonData
 }
